:sparkles: use tanstack retry to simplify auth checks
We can use the retry to log people out. It also affords some retries to prevent logging them out with a well timed refresh.

Also pocketbase's auth model includes custom fields, so I don't need to fetch the user. Unsure if this is a new feature or a previous misunderstanding.